fix: rename repository.interface -> interface
