+++ src/domain/interface/analytics.interface.ts
