+++ src/domain/interface/auth.interface.ts
