+++ src/domain/interface/communication.interface.ts
