+++ src/domain/interface/file.interface.ts
