+++ src/domain/interface/location.interface.ts
