+++ src/domain/interface/notification.interface.ts
