+++ src/domain/interface/operations.interface.ts
