+++ src/domain/interface/organization.interface.ts
