+++ src/domain/interface/user.interface.ts
