+++ src/domain/services/user.service.ts
@@ -1,7 +1,7 @@
 // src/domain/services/user.service.ts
 import { injectable, inject } from 'tsyringe';
 import { TOKENS } from '../../core/di/tokens';
-import type { IUserRepository } from '../repositories/user.repository.interface';
+import type { IUserRepository } from '../interface/user.interface';
 import type { CreateUserInput, UpdateUserInput, User } from '../entities';
 import type { AsyncResult } from '../../shared/response';
 import {
